+++ 06-build-page/index.js
@@ -65,6 +65,7 @@ const assetsSourcePath = resolveRootPath('assets');
 
 async function copyFolder(from, to) {
   try {
+
     await fs.rm(to, {
       recursive: true,
       force: true,
@@ -82,7 +83,11 @@ async function copyFolder(from, to) {
       const srcPath = path.resolve(from, entry.name);
       const destPath = path.resolve(to, entry.name);
 
-      await fs.copyFile(srcPath, destPath);
+      if (entry.isFile()) {
+        await fs.copyFile(srcPath, destPath);
+      } else if (entry.isDirectory()) {
+        await copyFolder(srcPath, destPath);
+      }
     }
   } catch (error) {
     console.error('Error copying folder:', error);
